requirements.txt
beautifulsoup4==4.11.1
blis==0.7.7
breadability==0.1.20
catalogue==2.0.7
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.1.0
cssselect==1.1.0
cymem==2.0.6
dask==2022.5.2
docopt==0.6.2
et-xmlfile==1.1.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.7.1
fsspec==2022.5.0
idna==3.3
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.1.0
langcodes==3.3.0
locket==1.0.0
lxml==4.9.0
MarkupSafe==2.1.1
murmurhash==1.0.7
newspaper3k==0.2.8
nltk==3.7
numpy==1.22.4
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
partd==1.2.0
pathy==0.6.1
Pillow==9.1.1
preshed==3.0.6
prettytable==3.3.0
pycountry==22.3.5
pydantic==1.8.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.6.2
requests==2.27.1
requests-file==1.5.1
scikit-learn==1.1.1
scipy==1.8.1
sgmllib3k==1.0.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
sumy==0.10.0
tabulate==0.8.9
thinc==8.0.17
threadpoolctl==3.1.0
tinysegmenter==0.3
tldextract==3.3.0
toolz==0.11.2
tqdm==4.64.0
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
wasabi==0.9.1
wcwidth==0.2.5
XlsxWriter==3.0.3
xmltodict==0.13.0
